feat: merge api
