--- cmd/followapi/Dockerfile
@@ -1,21 +1,0 @@
-FROM golang:1.20-alpine AS builder
-WORKDIR /app
-ENV CGO_ENABLED 0
-#ENV GOPROXY https://goproxy.cn,direct
-#RUN sed -i 's/dl-cdn.alpinelinux.org/mirrors.ustc.edu.cn/g' /etc/apk/repositories
-
-RUN apk update --no-cache && apk add --no-cache tzdata
-COPY . .
-RUN go mod download
-RUN go build -ldflags="-s -w" -o /app/followapi
-
-FROM scratch
-COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
-COPY --from=builder /usr/share/zoneinfo/Asia/Shanghai /usr/share/zoneinfo/Asia/Shanghai
-ENV TZ Asia/Shanghai
-WORKDIR /app
-COPY --from=builder /app/followapi /app/followapi
-# configs
-COPY configs /app/configs
-EXPOSE 8080
-CMD ["./followapi"]
