+++ cmd/userapi/api/api.go
@@ -14,12 +14,16 @@
 package api
 
 import (
+	"strconv"
+
 	"github.com/labstack/echo/v4"
 	"github.com/minio/minio-go/v7"
+	"github.com/sixwaaaay/sharing/pkg/encoder"
 	"github.com/sixwaaaay/sharing/pkg/pb"
 	"github.com/sixwaaaay/sharing/pkg/sign"
 )
 
+// UserApi defines '/users' api
 type UserApi struct {
 	mc     *minio.Client
 	uc     pb.UserServiceClient
@@ -39,4 +43,64 @@ func NewUserApi(mc *minio.Client, uc pb.UserServiceClient, bucket string, secret
 func (u *UserApi) Update(e *echo.Echo) {
 	e.GET("/users/:id", u.Profile, echo.WrapMiddleware(sign.Middleware(u.secret, false)))
 	e.PATCH("/users", u.UpdateProfile, echo.WrapMiddleware(sign.Middleware(u.secret, true)))
+}
+
+// FollowApi defines '/follow' api
+type FollowApi struct {
+	uc     pb.UserServiceClient
+	secret []byte
+}
+
+func NewFollowApi(uc pb.UserServiceClient, secret string) *FollowApi {
+	return &FollowApi{
+		uc:     uc,
+		secret: []byte(secret),
+	}
+}
+
+func (f *FollowApi) Update(e *echo.Echo) {
+	e.POST("/follow/following", f.Following, echo.WrapMiddleware(sign.Middleware(f.secret, false)))
+	e.POST("/follow/followers", f.Followers, echo.WrapMiddleware(sign.Middleware(f.secret, false)))
+	e.POST("/follow", f.Follow, echo.WrapMiddleware(sign.Middleware(f.secret, true)))
+	e.POST("/follow/friends", f.Friends, echo.WrapMiddleware(sign.Middleware(f.secret, true)))
+}
+
+func (f *FollowApi) subjectId(ctx echo.Context) (int64, error) {
+	subjectID, _ := ctx.Request().Context().Value("x-id").(string)
+	id, err := strconv.ParseInt(subjectID, 10, 64)
+	return id, err
+}
+
+func (f *FollowApi) Follow(c echo.Context) error {
+	var req pb.FollowActionRequest
+	if err := encoder.Unmarshal(c.Request().Body, &req); err != nil {
+		return echo.NewHTTPError(403, "invalid request")
+	}
+	id, err := f.subjectId(c)
+	if err != nil {
+		return echo.NewHTTPError(403, "invalid token")
+	}
+	req.SubjectId = id
+	reply, err := f.uc.FollowAction(c.Request().Context(), &req)
+	if err != nil {
+		return echo.NewHTTPError(403, err.Error())
+	}
+	return encoder.Marshal(c.Response().Writer, reply)
+}
+
+func (f *FollowApi) Friends(c echo.Context) error {
+	var req pb.GetFriendsRequest
+	if err := encoder.Unmarshal(c.Request().Body, &req); err != nil {
+		return echo.NewHTTPError(403, "invalid request")
+	}
+	id, err := f.subjectId(c)
+	if err != nil {
+		return echo.NewHTTPError(403, "invalid token")
+	}
+	req.SubjectId = id
+	reply, err := f.uc.GetFriends(c.Request().Context(), &req)
+	if err != nil {
+		return echo.NewHTTPError(403, err.Error())
+	}
+	return encoder.Marshal(c.Response().Writer, reply)
 }
