+++ readme.md
@@ -11,7 +11,6 @@ sharing is a short-form, video-sharing app backend
 [![userapi-ci](https://github.com/sixwaaaay/sharing/actions/workflows/userapi-ci.yaml/badge.svg)](https://github.com/sixwaaaay/sharing/actions/workflows/userapi-ci.yaml)
 [![commentapi-ci](https://github.com/sixwaaaay/sharing/actions/workflows/commentapi-ci.yaml/badge.svg)](https://github.com/sixwaaaay/sharing/actions/workflows/commentapi-ci.yaml)
 [![videoapi-ci](https://github.com/sixwaaaay/sharing/actions/workflows/videoapi-ci.yaml/badge.svg)](https://github.com/sixwaaaay/sharing/actions/workflows/videoapi-ci.yaml)
-[![followapi-ci](https://github.com/sixwaaaay/sharing/actions/workflows/followapi-ci.yaml/badge.svg)](https://github.com/sixwaaaay/sharing/actions/workflows/followapi-ci.yaml)
 [![account-ci](https://github.com/sixwaaaay/sharing/actions/workflows/account-ci.yaml/badge.svg)](https://github.com/sixwaaaay/sharing/actions/workflows/account-ci.yaml)
 
 ## License
